html: work around bug in html generation
Using texlive on Fedora 21, \url for some reason replaces
hypen - with -_ in html output only.

This is only a formatting issue: the link itself is correct.
For example:
<A HREF="http://docs.oasis-open.org">
http://docs.oasis-_open.org</A>

I don't know why this happens, but it seems to go away if we
override - within urls inserting an mbox around it.
makeatletter \g@addto@macro\UrlSpecials{\do\-{\mbox{-}}}

Signed-off-by: Michael S. Tsirkin <mst@redhat.com>

git-svn-id: https://tools.oasis-open.org/version-control/svn/virtio@474 0c8fb4dd-22a2-4bb5-bc14-6c75a5f43652