vsock: add vsock device
The virtio vsock device is a zero-configuration socket communications
device.  It is designed as a guest<->host management channel suitable
for communicating with guest agents.

vsock is designed with the sockets API in mind and the driver is
typically implemented as an address family (at the same level as
AF_INET).  Applications written for the sockets API can be ported with
minimal changes (similar amount of effort as adding IPv6 support to an
IPv4 application).

Unlike the existing console device, which is also used for guest<->host
communication, multiple clients can connect to a server at the same time
over vsock.  This limitation requires console-based users to arbitrate
access through a single client.  In vsock they can connect directly and
do not have to synchronize with each other.

Unlike network devices, no configuration is necessary because the device
comes with its address in the configuration space.

The vsock device was prototyped by Gerd Hoffmann and Asias He.  I picked
the code and design up from them.

Cc: Gerd Hoffmann <kraxel@redhat.com>
Cc: Asias He <asias.hejun@gmail.com>
Cc: Michael S. Tsirkin <mst@redhat.com>
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>
---
v6:
 * Make CIDs 64-bits but reserve upper 32 bits for now [Michael]
 * Specify SHUTDOWN -> RST clean disconnect process [Ian]

v5:
 * Switch to new, unused Device ID 19 [Ian]
 * Drop unused ctrl virtqueue, no need to reserve last virtqueue [Ian]
 * Document that VIRTIO_VSOCK_OP_CREDIT_UPDATE packets are valid even if
   no VIRTIO_VSOCK_OP_CREDIT_REQUEST was previously received. [Ian]
 * Document that only payload bytes are counted for buffer space
   management, not header bytes [Ian]
 * List the reserved CIDs [Ian]

v4:
 * Add event virtqueue and "Device Events" device operation section that
   explains how transport reset works for migration.
 * Reorder virtqueues with rx/tx first, then ctrl/event (similar to
   virtio-net)
 * __le32/16 -> le32/16 for consistency with existing code snippets
 * Add missing conformance.tex subsections for socket device entry in
   table of contents

v3:
 * "VSock device" -> "Virtio socket device" in free text [Michael]
 * Extract normative statements and add references from conformance
   chapter [Michael]
v2:
 * Document guest_cid field
 * Use MAY/MUST/CAN according to RFC 2119
 * Remove datagram socket type for the time being.  This can be added in
   the future but there are currently no applications.
 * Drop 3-way handshake for stream sockets.  It is not needed since
   virtio-vsock is reliable, in-order delivery and spoofing source
   addresses is impossible.
 * Drop max_virtqueue_pairs configuration space field.  This field was
   never defined and Linux code does not support multiqueue.  It can be
   added back later, if necessary.