content: add virtio file system device
The work-in-progress virtio file system device transports Linux FUSE
requests between a FUSE daemon running on the host and the FUSE driver
inside the guest.

This is an early version of the spec that maps FUSE requests to
virtqueues.  No changes are needed to the FUSE request format.

Multiqueue is supported for normal requests.  FUSE_INTERRUPT and
FUSE_FORGET requests are only sent on the dedicated hiprio queue.
Notifications are sent on the notifications queue.

The FUSE driver currently works in a "pull" model where userspace reads
requests from /dev/fuse one at a time.  Virtqueues are a "push" model
where the FUSE driver will need to enqueue requests onto a specific
virtqueue and wait for the guest to process them.

The request queue buffers are completed by the device when the request
has been processed and struct fuse_out_header has been filled out.  The
FUSE driver then picks up the completed request and processes it as if
the FUSE daemon had written to /dev/fuse.

Notifications involve device-to-driver communication.  Since virtqueues
live in guest RAM, the device cannot initiate communication.  Instead
the notifications queue is populated with empty buffers by the FUSE
driver (similar to a NIC rx queue).  The device then "completes" a
buffer when it wishes to notify the driver.  Replies to the notification
are place in a normal request queue, they do not go via the
notifications queue.

Note that this design assumes that the driver knows the required buffer
size for each request.  My understanding is that this is true in FUSE.
The only exception is FUSE_NOTIFY_STORE, and even there the FUSE
implementation has a limit of 32 pages, which makes for a natural buffer
size limit for the notifications queue.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>